Refactor visual seeding helpers
