+++ tests/visual/game-session.spec.ts
@@ -1,6 +1,14 @@
 import { test, expect } from '@playwright/test';
 import { waitForContentStable, hideDynamicContent, waitForInteraction } from './utils/wait-helpers';
 import { seedTestData, mockApiEndpoints } from './utils/data-seeder';
+import {
+  ensureSuggestedActionsExpanded,
+  renderSeededSuggestedActions,
+  seedInventoryItemsForVisual,
+  ensureSuggestedActionsContentVisible,
+  removeDuplicateSuggestedActionsTextarea,
+  relaxStoryColumnHeight,
+} from './utils/game-session-page-seeder';
 
 /**
  * Game Session Visual Regression Tests
@@ -353,358 +361,23 @@ test.describe('Game Session Visual Tests', () => {
     });
     console.log('🎨 Fade overlay:', fadeOverlay);
 
-    // Ensure Suggested Actions are expanded so the screenshot captures AI recommendations
-    const suggestedActionsTitle = page.locator('[data-testid="collapsible-section-title"]', {
-      hasText: 'Suggested Actions'
-    });
+    await ensureSuggestedActionsExpanded(page);
+    await renderSeededSuggestedActions(page);
+    await seedInventoryItemsForVisual(page);
 
-    if (await suggestedActionsTitle.count()) {
-      const toggle = suggestedActionsTitle.first().locator('..').locator('..').locator('[data-testid="collapsible-section-toggle"]');
-      const isExpanded = await toggle.getAttribute('aria-expanded');
-      if (isExpanded !== 'true') {
-        await toggle.click();
-        await page.waitForTimeout(200);
-      }
-
-      await page.evaluate(() => {
-        const title = Array.from(document.querySelectorAll('[data-testid="collapsible-section-title"]'))
-          .find((el) => el.textContent?.includes('Suggested Actions'));
-        if (!title) return;
-        const sectionEl = title.closest('[data-testid="collapsible-section"]');
-        if (!sectionEl) return;
-        const content = sectionEl.querySelector('[data-testid="collapsible-section-content"]');
-        if (content) {
-          const element = content as HTMLElement;
-          element.classList.add('block');
-          element.classList.remove('hidden');
-          element.setAttribute('aria-hidden', 'false');
-          element.style.display = 'block';
-          element.style.maxHeight = 'none';
-
-          const testWindow = window as typeof window & {
-            __TEST_DECISIONS__?: Record<string, {
-              prompt?: string;
-              options?: Array<{
-                id: string;
-                text: string;
-                hint?: string;
-                alignment?: string;
-                skillRequirements?: Array<{
-                  skillName?: string;
-                  targetId?: string;
-                  operator?: string;
-                  value?: number | string;
-                }>;
-                requirements?: Array<{
-                  type?: string;
-                  targetId?: string;
-                  operator?: string;
-                  value?: number | string;
-                  skillName?: string;
-                }>;
-              }>;
-            }>;
-            useInventoryStore?: {
-              setState: (
-                partial: unknown,
-                replace?: boolean
-              ) => void;
-              getState?: () => {
-                items: Record<string, unknown>;
-                entities?: Record<string, unknown>;
-                characterInventories: Record<string, string[]>;
-              };
-            };
-          };
-
-          const seededDecision = testWindow.__TEST_DECISIONS__?.['decision-cyberpunk-route'];
-          const skillNameLookup: Record<string, string> = {
-            'skill-hacking': 'Hacking',
-            'skill-streetwise': 'Streetwise',
-          };
-
-          const formatSkillRequirement = (requirement: any) => {
-            if (!requirement) return 'Skill Requirement';
-
-            const operatorSuffix = (operator: string, value: unknown) => {
-              if (value === undefined || value === null || value === '') return '';
-              const numericValue = typeof value === 'number' || typeof value === 'string' ? value : '';
-
-              switch (operator) {
-                case 'gte':
-                  return `${numericValue}+`;
-                case 'gt':
-                  return `>${numericValue}`;
-                case 'lte':
-                  return `≤${numericValue}`;
-                case 'lt':
-                  return `<${numericValue}`;
-                case 'eq':
-                  return `=${numericValue}`;
-                case 'neq':
-                  return `≠${numericValue}`;
-                default:
-                  return `${numericValue}`;
-              }
-            };
-
-            const baseName =
-              requirement.skillName ||
-              skillNameLookup[requirement.targetId as string] ||
-              requirement.targetId ||
-              'Skill Requirement';
-
-            const suffix = operatorSuffix(requirement.operator, requirement.value);
-            return suffix ? `${baseName} • ${suffix}` : baseName;
-          };
-
-          if (seededDecision?.options?.length) {
-            const badgeClasses = [
-              'inline-flex items-center rounded-full px-2.5 py-0.5 text-xs font-semibold',
-              'transition-colors focus:outline-none focus:ring-2 focus:ring-blue-500 focus:ring-offset-2',
-              'border border-gray-300 bg-gray-100 text-gray-700',
-            ].join(' ');
-
-            const optionMarkup = seededDecision.options
-              .map((option) => {
-                const resolvedSkillRequirements = (() => {
-                  if (Array.isArray(option.skillRequirements) && option.skillRequirements.length > 0) {
-                    return option.skillRequirements.map((req) =>
-                      req?.skillName || req?.label || 'Skill Requirement'
-                    );
-                  }
-
-                  if (Array.isArray(option.requirements)) {
-                    const rawSkillRequirements = option.requirements.filter(
-                      (req: any) => req?.type === 'skill'
-                    );
-                    if (rawSkillRequirements.length > 0) {
-                      return rawSkillRequirements.map((req: any) => formatSkillRequirement(req));
-                    }
-                  }
-
-                  return [];
-                })();
-
-                const skillMarkup = resolvedSkillRequirements.length
-                  ? `<div class="flex flex-wrap gap-1 mt-2" role="group" aria-label="Skill requirements">
-                      ${resolvedSkillRequirements
-                        .map(
-                          (label: string, index: number) => `
-                          <div class="${badgeClasses}" data-testid="skill-badge-${option.id}-${index}">
-                            ${label}
-                          </div>
-                        `
-                        )
-                        .join('')}
-                    </div>`
-                  : '';
-
-                const optionClasses = [
-                  'block w-full text-left p-3 border rounded-md shadow-sm mb-2 transition-colors',
-                  'bg-white hover:bg-gray-50 focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-blue-500 focus-visible:ring-offset-2',
-                ].join(' ');
-
-                return `
-                  <button class="${optionClasses}" data-testid="choice-option-${option.id}">
-                    <div class="font-medium text-gray-900">${option.text}</div>
-                    ${option.hint ? `<div class="text-sm text-gray-500 mt-1">${option.hint}</div>` : ''}
-                    ${skillMarkup}
-                  </button>
-                `;
-              })
-              .join('');
-
-            element.innerHTML = `
-              <div class="space-y-2" role="radiogroup" aria-labelledby="choices-heading">
-                ${optionMarkup}
-              </div>
-            `;
-          }
-        }
-        const toggleEl = sectionEl.querySelector('[data-testid="collapsible-section-toggle"]');
-        if (toggleEl) {
-          toggleEl.setAttribute('aria-expanded', 'true');
-          toggleEl.textContent = '−';
-        }
-
-        // Seed inventory with demo items so the visual baseline reflects populated inventory UI
-        const inventoryStore = (window as typeof window & {
-          useInventoryStore?: {
-            setState: (
-              partial: unknown,
-              replace?: boolean
-            ) => void;
-            getState?: () => {
-              items: Record<string, unknown>;
-              entities?: Record<string, unknown>;
-              characterInventories: Record<string, string[]>;
-            };
-          };
-        }).useInventoryStore;
-
-        if (inventoryStore?.setState) {
-          const cyberpunkItems = {
-            'inventory-ghostlink-cyberdeck': {
-              id: 'inventory-ghostlink-cyberdeck',
-              name: 'Ghostlink Cyberdeck',
-              description: 'Signature deck tuned to slip past Arasaka intrusion countermeasures.',
-              categoryId: 'equipment',
-              quantity: 1,
-              stackable: false,
-              acquisitionHistory: [
-                {
-                  acquiredAt: '2024-01-01T01:58:00.000Z',
-                  method: 'quest',
-                  quantity: 1,
-                  description: 'Recovered during the vault raid briefing.',
-                },
-              ],
-              categorization: {
-                categoryId: 'equipment',
-                source: 'manual',
-                classifiedAt: '2024-01-01T01:58:00.000Z',
-              },
-              createdAt: '2024-01-01T01:58:00.000Z',
-              updatedAt: '2024-01-01T02:04:00.000Z',
-            },
-            'inventory-neuro-stims': {
-              id: 'inventory-neuro-stims',
-              name: 'NeuroBoost Stims',
-              description: 'Fast-acting injectors that keep reflexes sharp during breach attempts.',
-              categoryId: 'consumables',
-              quantity: 3,
-              stackable: true,
-              maxStack: 5,
-              acquisitionHistory: [
-                {
-                  acquiredAt: '2024-01-01T01:45:00.000Z',
-                  method: 'purchase',
-                  quantity: 3,
-                  description: 'Purchased from a trusted ripperdoc contact.',
-                },
-              ],
-              categorization: {
-                categoryId: 'consumables',
-                source: 'manual',
-                classifiedAt: '2024-01-01T01:45:00.000Z',
-              },
-              createdAt: '2024-01-01T01:45:00.000Z',
-              updatedAt: '2024-01-01T02:02:30.000Z',
-            },
-            'inventory-black-ice-shard': {
-              id: 'inventory-black-ice-shard',
-              name: 'Black ICE Shard',
-              description: 'Prototype defensive program that can be slotted into the deck on demand.',
-              categoryId: 'quest-items',
-              quantity: 1,
-              stackable: false,
-              acquisitionHistory: [
-                {
-                  acquiredAt: '2024-01-01T02:01:00.000Z',
-                  method: 'reward',
-                  quantity: 1,
-                  description: 'Rewarded by Fixer Nyx after completing the reconnaissance run.',
-                },
-              ],
-              categorization: {
-                categoryId: 'quest-items',
-                source: 'manual',
-                classifiedAt: '2024-01-01T02:01:00.000Z',
-              },
-              createdAt: '2024-01-01T02:01:00.000Z',
-              updatedAt: '2024-01-01T02:03:10.000Z',
-            },
-          } as Record<string, unknown>;
-
-          inventoryStore.setState((state: {
-            items: Record<string, unknown>;
-            entities?: Record<string, unknown>;
-            characterInventories: Record<string, string[]>;
-          }) => ({
-            items: {
-              ...state?.items,
-              ...cyberpunkItems,
-            },
-            entities: {
-              ...state?.items,
-              ...cyberpunkItems,
-            },
-            characterInventories: {
-              ...state?.characterInventories,
-              'char-cyberpunk-hacker': [
-                'inventory-ghostlink-cyberdeck',
-                'inventory-neuro-stims',
-                'inventory-black-ice-shard',
-              ],
-            },
-          }));
-        }
-      });
-
-      await page.waitForSelector('[data-testid^="choice-option-"]', {
-        state: 'visible',
-        timeout: 2000,
-      });
-
-      await page.waitForSelector('.inventory-item', {
-        state: 'visible',
-        timeout: 2000,
-      });
-    }
-
-    // Force the collapsible section content to be visible (defensive in case layout toggles collapse again)
-    await page.evaluate(() => {
-      const content = document.querySelector('[data-testid="collapsible-section-content"]');
-      if (!content) return;
-      const element = content as HTMLElement;
-      element.classList.add('block');
-      element.classList.remove('hidden');
-      element.setAttribute('aria-hidden', 'false');
-      element.style.display = 'block';
-      element.style.maxHeight = 'none';
-      element.style.opacity = '1';
-      const toggle = document.querySelector('[data-testid="collapsible-section-toggle"]');
-      if (toggle) {
-        toggle.setAttribute('aria-expanded', 'true');
-      }
-    });
-
-    // Wait for at least one suggested action button to be visible in the DOM
     await page.waitForSelector('[data-testid^="choice-option-"]', {
+      state: 'visible',
       timeout: 2000,
-      state: 'visible'
     });
 
-    // Remove duplicate textarea that can appear inside the suggested actions section in tests
-    await page.evaluate(() => {
-      const suggestedActionsSection = document.querySelector('[data-testid="collapsible-section"]');
-      if (!suggestedActionsSection) return;
-
-      const duplicateTextarea = suggestedActionsSection.querySelector('textarea');
-      if (!duplicateTextarea) return;
-
-      const wrapper = duplicateTextarea.closest('.mb-4, .p-4, .bg-gray-100');
-      if (wrapper && suggestedActionsSection.contains(wrapper)) {
-        wrapper.remove();
-        console.log('✅ Removed duplicate textarea wrapper inside collapsible section');
-        return;
-      }
-
-      duplicateTextarea.remove();
-      console.log('✅ Removed duplicate textarea inside collapsible section');
+    await page.waitForSelector('.inventory-item', {
+      state: 'visible',
+      timeout: 2000,
     });
 
-    // Temporarily remove height constraints to show all segments in fullPage screenshot
-    await page.evaluate(() => {
-      const storyColumn = document.querySelector('.lg\\:flex-1.min-h-0.flex.flex-col.lg\\:overflow-hidden.relative');
-      if (storyColumn) {
-        const element = storyColumn as HTMLElement;
-        element.style.maxHeight = 'none';
-        element.style.overflow = 'visible';
-        console.log('✅ Removed height constraints for fullPage screenshot');
-      }
-    });
+    await ensureSuggestedActionsContentVisible(page);
+    await removeDuplicateSuggestedActionsTextarea(page);
+    await relaxStoryColumnHeight(page);
 
     await page.waitForTimeout(200); // Wait for layout to adjust
